i
Former-commit-id: da442d882d298372e82f923faef975a6f1193bc8

+++ crates/brontes-inspect/src/composer/utils.rs
@@ -62,8 +62,8 @@ pub(crate) fn build_mev_header(
         .sum::<u128>();
 
     //TODO: need to check if decimals are correct
-    let builder_eth_profit = (total_bribe + pre_processing.cumulative_gas_paid
-        - metadata.proposer_mev_reward.unwrap_or_default()) as i128;
+    let builder_eth_profit = (total_bribe as i128 + pre_processing.cumulative_gas_paid as i128)
+        - (metadata.proposer_mev_reward.unwrap_or_default() as i128);
 
     MevBlock {
         block_hash: pre_processing.meta_data.block_hash.into(),
